shortcuts: combine fsmaps and shortcuts into just shortcuts
It's a pain to maintain both fsmaps and shortcuts seperate from each
other but within the same directory. Now we maintain them both in the
same file.

Because of this, the directory hirearchy has been trashed, don't need to
have subdirectories for a single file :P. Instead the default shortcuts
for a platform are in a file named after that platform.

.bash_aliases also sources shortcuts.private and $platform.private if
they exist, allowing you to define host local shortcuts. :smile:.