plugins/subbot: always run from same directory as config file
you're tacitly allowed to specify config files in upto 3 locations,
yet the subbot is always invoked from the same path (a directory
matching the path argument), but this may not exist and more
importantly it makes writing configs cumbersome. You write as if the
cwd is different to the one your config is in.

Now your subbots are always invoked from the same directory as the
config file.