misc: switch from cygwin to msys
msys is a windows alternative to cygwin, that instead of replacing the
windows filesystem with a unix like filesystem, tries to merge the two
in a sensible way.

I got tired of cygwin demanding I install special alternatives to
programs which work well for them. It's just too much of a pain to
interop between the scores of windows programs and the handful of
cygwin ones.