programs/fzf: add package + new command fzf-dmenu
fzf is the interactive fuzzy finder. it let's you incrementally select
from a large list of options.

This commit adds fzf to the list of optional install targets for arch
and also introduces a new script fzf-dmenu which acts as a dmenu
alternative. It lets you select a program to run using a fzf popup.