emacs-focus: support focusing between terminal and gui emacs
This was a long time issue for me, my `emacs-focus` script only worked
with GUI frames because they have emacs as their main process. In the
terminal, the emacs instance is just a child process of the parent
terminal process.

I could've used the title of the emacs frame, but:
- st doesn't support title introspection commands, so `xterm-title.el`
  doesn't work.
- I'd need to choose a title format that no non-emacs window could
  possibly have. If I'm googling for emacs then chrome/firefox/etc
  update their titles to have emacs in it, meaning `emacs-focus` would
  take me to my browser... completely defeating the point of it.

Seeing as I only use emacs from within tmux (because of no native
support for st) I've configured tmux to update st's title to show
the name of the current window; which by default is the name of the
command in the active pane for that window.

I've also refactored my `app-windows` script to allow you to return a
list of windows which by their window title.

Lastly I've modified both emacs and tmux to include my hostname in
their titles, to raise the accuracy of window searches.