deploy: (shenv): defer platform check until after path update
Okay, this was a strange bug to track down. I'm using a script local to
my home directory to find out the current platform and incorporating that
in a bunch of places. However I was updating my PATH variable to include
where that script can be found AFTER already calling it.

In conventional usage this wasn't a problem because my programs we're
inheriting their PATH from some other program in my user group so it'll
already have had the correct path, but for login shells and such this
wasn't the case and an error was written to stderr.

This commit pushes the platform check until after the PATH update. fdaedf01b5b1d7da20a2117ccbe7d981f3c7a974