(preview): cleanup script
refactored file/mime specific previewers into bash functions which are
then invoked from different handlers.

This was necessary because files with a JSON mimetype weren't being
previewed as JSON (we were only using extensions for that). This is
also a more flexible approach compared to before.

Finally I've removed the weird behaviour around exit codes. Now if the
program managed to preview the way it wanted to it exits with 0.
Otherwise it keeps trying other possible preview methods until it
finds one or it echoes a placeholder preview and exits with a non-0 exit
code. The 2 exit code which was supposed to mean just cat this file
has now been replaced, instead this script will cat the file itself
and exit with a 0 exit code.