(desktop): remap caps-lock to control
This has been in my config for a while but I wanted to wait to commit
it until after I found a way to get the same affect on windows. Lo and
behold I finally did it so here it is.

The rationale for this was pretty simple. Caps-Lock with my bluetooth
keyboard SUCKS!!! See [[https://bugs.launchpad.net/ubuntu/+source/bluez/+bug/1341256][this]] 6 year old issue for an explanation as to
why. Anyways I got tired of it my keyboard becoming useless everytime
I hit caps-lock so I remapped it to something else in the hopes that
that'll fix it. I could've gone with escape or control and in the end
I settled on control because it made hitting emacs modifier keys much
less stressful on the fingers.

This commit:
- uses setxkbmap to remap caps-lock to control on linux systems.
- uses a regedit to do the same on windows platforms. Like with
  anything useful on windows, doing so requires elevated privilages and
  windows won't bother asking you for them when their required; so run
  from an admin console lest you beware.