(bin): add xrun script to execute desktop files
xrun just imports xopen and uses its existing lookup and execution
routines to let you directly execute a Desktop file. I've modified
fzf-dmenu to use this instead of trying to parse out and format the
Exec line itself (either way I've traded one python call for another,
so there shouldn't be much of a difference).

I've also modified fzf-dmenu to present both the name of the program
being executed and the desktop file from which it was recieved in a
tabulated list form. This is because the same program may be present
from multiple desktop entries and its confusing trying to choose one
without it.

Also added an alias for fzf-dmenu directly from your shell.