(tmux): add main tmuxinator config
I've developed a tiny habit of creating a new tmux session, renaming
it main and then creating the same program in the first/second windows
while using the rest for random things.

Now I'm automating all of that with tmuxinator.

For this config I use the first window as a dashboard to show system
information throughout my hosts lifetime. The second window monitors
network traffic. If the programs needed for these windows don't exist
then it tries some fallbacks and eventually just kills the window when
there isn't anything on the system to show the information I wanted.

The last window is just a shell.