(tmux): replace fzf-tmux with tmux-popup
A refactor of fzf-tmux designed to support the new popup feature in tmux. It can be
used to run arbitrary interactive commands (which is why it lacks the fzf prefix).

I've refactored most of my fzf-tmux scripts to use this now and I've cleaned up my
bindings for these scripts in tmuxrc.