(bin): add nfc
A numbered-folder-constructor script. Takes a template directory and then creates a
bunch of new folders from it in a number range. A nice way to create structured
folders without having to manually renumber them after the fact.

This script automatically detects where a number occurs in an input folder and
incrementally produces output folders.