(preview): add option to include directory
By default when trying to preview directories only the contents of the directory is
shown. This makes it hard to track where the preview came from when the associated
path doesn't include it.

Now if the INCLUDE_DIR options is set to 1, preview will also add a header which
includes the path to the cwd (using ls to maintain LS_COLORS) and then a seperator
which is just a series of dashes of the same length as the path.