(ls-repos): allow direct repo path in REPO_PATH
Before REPO_PATH only pointed to directories that contain repositories, this didn't
scale well when met with repos outside of collections of repositories. For example,
my dotfiles are at ~/.dotfiles/, but no other directory in ~/ points to a repository
so checking them for it by putting ~/ into REPO_PATH wasn't very practical.

This patch alters REPO_PATH to allow you to specify direct paths to repositories. Now
a trailing slash for any path in REPO_PATH indicates that path is a container for
multiple repositories. A lack of a slash indicates that path is a repository itself.

There's a slight performance cost due to this, but flexibility always has some cost
🕹.