(shells): add exit-code info to prompt
Adapted my error-colors script to an exit_color autoload which is used to color and
include the exit-code in my shell prompt. For zsh this is to the right, just before
the user/host section (only when there's a none-0 exit code). For bash it's just
before the git info.