(bin): ls-exec prevent duplicates
So in shenv, I'm assigning the PATH variable for my current shell by preppending a
few paths before the existing PATH variable value. Now my shell is spawned as a child
of the system shell (which also uses this same shenv config) so I end up with a lot
of duplicate entries in my PATH.

ls-exec then compounds this issue by not disregarding duplicate paths, meaning if
~/.local/bin appears in your PATH twice and foo is at this path, ls-exec outputs foo
twice even though it points to the exact same executable. Now I use an awk cache to
strip out duplicate paths as I encounter them. This adds a slight startup overhead
but I consider it negligable.

I've also added support for using GNU parallel instead of xargs (when it's available
on your system). This approach will spawn less find processes... but parallel will
wait until it recieves enough input before starting the first process. So for long
PATHs this'll be better in the long run, for short ones it slows startup.