(bin): add prog-icons
A tiny script to associate program names with unicode icons. I've been meaning to do
this for a while, but never knew when I'd ever really use it :/. But now it's here.

prog-icons reads a configuration file at ~/.config/progicons which follows the same
format as ~/.config/diricons. It can continuously read program names from STDIN or
those passed as positional arguments and then outpus the associated icon. There's a
forced fallback in place in case no pattern is matched against.

I've updated my tmux config to use this script to include a program icon (on the
right hand side). So far its working great. One thing to NOTE is that this program
tries to unbuffer the output stream when an unbuffering program exists. However
if it's not available, this script will probably never produce any output on the tmux
statusline. I suggest installing stdbuf (or in the case of cygwin which doesn't have
stdbuf, moving to linux).