(tmux): fix zombie tmux-icons and concurrent pane access issues
So my previous approach to icons on the status line had me creating a tmux-icons
process and having it continually query tmux for the current pane command which it
then output. A clear flaw with this approach was that only the current pane with
focus would have its icon shown. If I have two terminals each with their own tmux
sessions only the active (focused) one would be shown.

I tried to fix this by adding a -s flag to tmux-icons which accepted the current tmux
session. tmux-icons would then only query that session... but it turns out tmux spawns
a new process for each tmux client and doesn't kill them when the client is killed...
so I ended up leaving behind a tonne of tmux-icons processes.

Now I'm passing the tmux clients tty to the -s flag and I've set up tmux-icons to
crash as soon as a single process fails (most likely the process querying the current
session information). So if I kill a client connection... tmux-icons up and exits
alongside it. That should mean there's only as many tmux-icons processes as I have
tmux clients.