(bin): Add bang
I've already got a bang alias but it can only be used in interactive sessions,
situations such as in ranger/lf bell isn't available so I have to jump down to
a subshell. Now I've got a script that does the same thing as the bell alias. I'm
keeping the alias around just to avoid the overhead of spawning a subshell just to
echo a single character (probably negligable in practice).