(clipsink): Add flag to exit when stdin finishes
One minor annoyance with my previous clipsink implementation was that
the only way to terminate it was to C-c interrupt. That's good enough
but it doesn't work when you're in the middle of a pipeline, eg:

```sh
clipsink | awk -e '{ print "--url " $0 }' | curl --config -
```

Sending an interrupt crashes the entire pipeline, when what I really
want is to end the clipsink part so the rest of it can carry on working.

Well now with the -m flag clipsink will automatically exit when stdin
recieves EOF (eg. with C-d).