(tmux): Add disk space indicator
Inspired by the default [[https://github.com/polybar/polybar][polybar]] config, I've appended a section which
just shows various mount points and the amount (percentage) of disk-space
their taking up.

Mount points are specified in a passwd style file found at
$TMUX_HOME/disk-free.conf. The first column is the name as shown in the
status bar and the second column is where on the filesystem it can be
found. The default point is just the root of the file system.