(tmux): Add option to disk-free to show bytes
Instead of percentages you can now show a human readable indication of
the amount of free-space (eg. 10GB, 500MB). It's intentionally clamped
at a maximum of 3 digits, thus it can show 999B, but 1000B will be
shown as 0K (rounding down from 0.9KB). This will cycle upwards until we
reach a yotabyte.

To do so this patch adds a new library to my mohkale (bin/lib/mohkale)
package that provides support for converting to human readable values in
a series.