(tmux): Make theme cmds only output on change
If each of my tmux cmds (eg. disk-free or cpu-load) are about to output
something which matches it's exact previous output, then it won't output
anything.

The performance gains for this will probably be negligible, but there's
no need to output something and trigger a reparsing + restyling +
reprinting in tmux when the end result looks the same as it currently
does.