(tmux): Make tmux-icons subscribe to events
See [[https://github.com/tmux/tmux/issues/2598][tmux#2598]].

Now tmux-icons instead of periodically querying tmux through
display-message to detect when the active panes command has
changed (and thus an icon change is required), we instead subscribe
to an event that automatically outputs when this happens.

To do this I'm using the new [[https://github.com/tmux/tmux/wiki/Control-Mode][control-mode]] for tmux, which is still in
unreleased form, and basically creating a shadow client which follows
the same motions as my original client in a read-only non interactive
way.

The advantages of this are the same as before:
* You can have multiple clients in the same session and the icons for
  them will not interfere with each other.
* You can automatically get icon updates to match the active pane.

However the key advantages of this new approach over the previous one is:
* You don't have to keep polling for updates that may not have occurred.
* You're message-log isn't littered with useless display-message calls.
* You don't keep spawning tmux processes like a mad-man. Tmux is lightweight
  but the cost of continually making a new process again and again can add
  up considerably.

Implementation wise I'm using a co-process managed by gawk to kick-start
the shadow client and then continually monitor its output. When the target
client switches sessions the shadow client does so as well.
The program name (icon) will be dumped to stdout any other output from
the shadow-client can be found in stderr (for debugging purposes).