(xdg): Shorten downloads and documents directory
I got tired of having short 4/5 letter directories such as ~/media,
~/prog, ~/multi and then having downloads and documents which're almost
twice as long.

So now I've renamed:
- ~/downloads -> ~/files
- ~/documents -> ~/docs

~/files isn't exactly an ideal synonym for ~/downloads but it's brief,
and expresses the idea that this directory contains multiple files which
aren't exactly similar.
I did consider using fetch instead of files because it's used in a
similar context in git and expresses the intent well enough.

This commit also added two new dirs to xdgdirs (one for games and one
for local temp files) and had me restructure my alias configs to be a
little more readable.