(setup): Initial commit backup script
A script to scower the current host machine for any files you'd want to
keep after migrating to a new system and then copies those files over to
a backup directory.

Just a way to automate the ordeal of backups and to avoid the nightmare
of erasing your installation only to later recall some innocuous file
you need but ended up deleting.