(lf/tmux): Refactor fzf autoloads support
Essentially moved the _fzf() declaration for tmux and lf to what was the
select_path script (now renamed to setup-fzf).

I've also essentially cleaned up my entire lf config to be a lot less
repetitive and easier to jump towards. Bindings have been moved closer
to the commands that describe them and heading are now meaningful.