(tmux): Make tmux-popup not read/write by default
By default (unless stdin was a tty) tmux-popup cat-ed standard input
into a named pipe and then redirected that to the popup command before
reading the commands standard output and writing it out. This option
disables that feature and requires you to manually specify when you want
input and/or output redirection with the -i and -o flags.

This makes tmux-popup a generic commnd to spawn a command in a tmux
popup window or a new split and can work for innocuous commands such
as your EDITOR, your shell, or any other script.