(bin): Move evi from autoloads
evi is now officially a script again 🙋.

I was getting tired of keeping aliases both in my alias directory
~/.config/aliases and autoload related aliases in ~/.config/autoloads.
I did this mostly because aliases have this feature where if there's
a trailing slash in the alias it expands the second argument as an alias
as well. So for example you can alias `sudo ` as `sd` and `sd otheralias`
expands both sudo and your otheralias.
My alias script ended up unintentionally discarding these trailing
spaces because of how bashs read command is implemented. This commit
fixes that behaviour so that I can now include trailing whitespace in
aliases.

I've also update my evi script and select_path so that it works with evi
again. This was a laborious migration of several configuration
files (including with_emacs) but now you can use `we gx` again.