(tmux): Refactor theme cmds + fix timeout crashes
I wrote my theme cmds so that they wouldn't spawn too many processes or
re-acquire any pertinent meta-data more than once. The cost of spawning
a new process every second or so is generally negligable, but spawning,
processing and output a theme cmd every second (especially when in the
a lot of cases nothings changed from the previous print) was a jarring
inefficiency.

Unfortunately [[orgit:tmux][tmux]] itself has the behaviour of killing any status
processes that haven't outputted anything in over 60 minutes. This is
probably the safe move (the program is probably locked and killing and
restarting it will most likely force it to output something) however in
my case a lot of commands are assumed to not output anything for
prolonged periods (status-misc or disk-free). This commit introduces a
new MAX_OUTPUT_INTERVAL which forces the scripts to output something
when they approach this interval. This should at the very least prevent
any unintended crashes.

I've also refactored a lot of the repeated logic in the status cmds into
~shared.py~. It contains helpers to repeatedly print a cmd and other
useful goodies.