(prog): Add pass config
I've been using pass as my password manager for a while but I've never
moved it into my dotfiles because I was worried it wouldn't be secure to
have a link to it in a public repository. But I guess I was just being
paranoid. Everything is still GPG encrypted and now I don't have to hunt
down an SSH url every time I setup a new host and want my passwords
available.