(core): Add org directory as a sub-module
I've come to the realisation that dropbox isn't a reliable platform for
persistent notes. It works well enough but so much of it is hidden behind
paywalls or obfuscated APIs that're intentionally designed to be usable
only from within the Dropbox ecosystem. So for example earlier today I
used a batch `sed` process to try and change a link type in my org-brain.
I was worried my regexp might not've matched properly and so some
information was lost so I wanted to recover my previous changes. Turns
out dropbox allows you to recover a single file change under a free
plan, but not an entire directory change. So I can painstakingly go
through each file and remove the changes one by one (through an online
platform that seems allergic to any sort of automation in this regard)
or pay-up for an upgrade, make the change, and then promptly cancel my
subscription cause frankly £9.99 a month is an insane price tag. In the
end I had to pay to revert, but I could've just used a git-repo and not
had to deal with all this hassle.