(bin): Make app-windows/term-dwim respect KDE activities
One minor annoyance with the previous version of these scripts was that
when they would unintentionally change my current KDE activity because
X itself doesn't have the concept of activities and therefore sees all
windows in the same workspace as having the same precedence despite
being scattered across other activities.

This patch fixes that by using the _KDE_NET_WM_ACTIVITIES property on
windows to determine whether their in the current activity or not.
It detects whether we're in KDE or not using the DESKTOP_SESSION
environment variable.

This commit also adds some basic refactorings to both these scripts to
aid readability.