(tmux): Simplify tmux-icons into tmux-title
Now instead of subscribing to the current pane command and then piping
it into prog-icons, I do the piping from tmux directly and just use a
script to output the current panes cmd. This script is written in awk
as a simplification of tmux-icons.

From now on if I'm writing a long-lasting awk script I'll do it in a
dedicated file over embedded shell script.