(tmux): Prefer tmuxp to tmuxinator
I've been using tmuxinator for a while but it has a lot of dependencies
and is annoyingly slow at startup. tmuxp is a somewhat lighter alternative
that shares most of the same feature-set.

Firstly this converts any tmuxinator configs I have to tmuxp configs.

Secondly I've reworked fztmuxinator into fztt and refactored ttls to
support both tmuxinator and tmuxp configuration formats. The change
isn't all that substantial. I've mostly just copied over the previous
ttls into an inline script run in the new ttls. The difference now is
that it should be easy to add more layout scripts as and when I
encounter them.