(core): Remove ~/.config/prompt and ~/.config/bindings
Keeping this shell specific stuff separate from my shell was a weird
idiosyncrasy that seems completely unnecessary to me.
In retrospect I think it was just because bash doesn't have a
$XDG_CONFIG_HOME/bash directory and I didn't want to configure
to place these in the home directory so a grouping of them in
XDG_CONFIG_HOME made sense. However this approach is far more
reasonable.