(submodules): Move src submodules into dedicated subdirectories
I've been keeping the src repos for st and sxiv in awkward places. Sxiv
was in prog/sxiv/sxiv and st was in prog/st but the dotty configuration
for it was in prog/.config. Now when I have src submodules their placed
in the src subdirectory of where configurations for them would
ordinarily go.