(setup): Reimplement dotty in bash
I've been using my own dotty implemention in go for a while and did plan
to implement a full lisp runtime for it, but now that I've started a new
job and found that I don't have the discretion to pull in programs at
will, I've found myself at a loss. Bash is portable and reliable and
while slower, I very much doubt I'll ever be in a situation where it
isn't available. Because of this I've reimplemented dotty in bash.

I think it looks reasonably descriptive and atm there's feature parity
with the previous version. There's even some newer features such as
--dry-run which outputs a shell script listing out whatever the
installation would've run. Lastly this should also be easier to extend.
I can easily add new package managers or other extensions with little to
no hassle.