desktop: Add in progress wayland support
I won't deny, it's the future. I tried setting up a multi-monitor setup
with my NVidia graphics laptop and it was completely unusable. Xorg
couldn't adapt to the different display resolutions and refresh rates.
Wayland worked with no issue, out of the box.

I'm loving how well thought out and safe the API is. The only reason it
isn't enabled by default (atm) is because there's a bug with the
drag-and-drop functionality between xwayland and non-X apps with the KDE
compositor. Hopefully it'll be fixed soon.