tmux: Make status-misc asynchronous
* Reimplements all the status-misc segments as asynchronous functions
that can update themselves at distinct intervals. This provides much
more flexibility to this status-line segment. Elements that are liable
to change very frequently can be updated fast whereas elements that're
liable to not change for a long time can be deferred for longer periods.
As soon as any of them change, the output will be reprocessed and
printed.
* This also moves any shared status-line python functions into a lib
submodule for clarity.
* Passes all theme scripts through black and isort.