tmux: Fix bug in standalone status-misc segment scripts on py38
Turns out if you don't define a ChangeEvent object while in an event
loop, it'll keep requesting a new event loop and continually fail since
the original event loop waiting on it is separate to this new event
loop. To fix it I define the change event within the main event loop.

Note: This isn't an issue in py > 3.10