theme: Gracefully handle mullvad-vpn ip address resolution
Before I added a failover counter such that if the command succeeded it
would need to fail n times before the value is actually cleared. This is
proving to be more consistent than I expected. As a workaround we now
just ignore a failure to resolve the ip-address. If we cannot get one we
don't include it in the output. If we can then we do. This is configured
through 2 new options: --skip-ip and --ignore-missing-ip. The former
skips the ip-address resolution altogether. The latter just ignores it
if it couldn't be determined.