bin: Make ls-projects more friendly to hierarchies of project directories
Add a way to list projects by directory hierarchy in REPO_PATH over just the project
basename. This is useful for projects with very generic names like ui. Now you can
set a base directory for the project root in the REPO_PATH_ALIASES env variable and
instead ls-projects will show it as the path relative to that root or with a
substituted alias.

TL;DR; Now ls-projects shows ~/repos/NVChad/ui as NVChad/ui instead of just ui.